requirements.txt
absl-py==0.11.0
aiohttp==3.6.2
ale-py==0.7.1
appdirs==1.4.4
appnope==0.1.0
astroid==2.8.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.4
azure-core==1.18.0
azure-functions==1.7.2
azure-storage-blob==12.9.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.2
black==21.5b1
bs4==0.0.1
cachetools==4.1.1
camelcase==0.2
certifi==2020.11.8
cffi==1.14.6
chardet==3.0.4
click==7.1.2
cloudpickle==2.0.0
colorama==0.4.4
cryptography==3.4.8
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dnspython==1.16.0
filelock==3.0.12
flake8==3.9.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
gym==0.20.0
h5py==2.10.0
idna==2.9
imageio==2.10.1
importlib-resources==5.2.2
ipython==7.18.1
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
jedi==0.17.2
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.3.3
matplotlib==3.3.1
mccabe==0.6.1
more-itertools==8.2.0
msrest==0.6.21
multidict==4.7.5
mypy-extensions==0.4.3
mysql-connector-python==8.0.19
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==20.3
pandas==1.1.2
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2021.5.29
pipfile==0.0.2
platformdirs==2.4.0
pluggy==0.13.1
polyline==1.4.0
prompt-toolkit==3.0.7
protobuf==3.14.0
ptyprocess==0.6.0
py==1.8.1
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyglet==1.5.21
Pygments==2.6.1
pylint==2.11.1
pyparsing==2.4.6
pypolyline==0.2.5
pyspark==3.1.1
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.11.13
reprint==0.6.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
stable-baselines==2.10.2
summa==1.2.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.52.0
traitlets==5.0.2
typing-extensions==3.10.0.2
urllib3==1.25.8
virtualenv==20.1.0
virtualenv-clone==0.5.7
wcwidth==0.1.9
websockets==8.1
Werkzeug==1.0.1
wrapt==1.11.2
yapf==0.31.0
yarl==1.4.2
zipp==3.6.0
